+++ ios/Mobile/CODocument.mm
@@ -84,6 +84,9 @@ static std::atomic<unsigned> appDocIdCounter(1);
                                [NSURLQueryItem queryItemWithName:@"lang" value:app_locale],
                                [NSURLQueryItem queryItemWithName:@"appdocid" value:[NSString stringWithFormat:@"%u", appDocId]],
                                [NSURLQueryItem queryItemWithName:@"userinterfacemode" value:([[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPad ? @"notebookbar" : @"classic")],
+                               // Related to issue #5841: the iOS app sets the
+                               // base text direction via the "dir" parameter
+                               [NSURLQueryItem queryItemWithName:@"dir" value:app_text_direction],
                              ];
 
     NSURLRequest *request = [[NSURLRequest alloc]initWithURL:components.URL];
